Added options for Dice so Nice Customization
As it is, the dice will pretty much ignore any changes you make inside of Dice so Nice, for ease of readability. And that's awesome if your players doesn't really care about flourishing things a lot.

If they do, however, having the same black and red or orange dice all the time, without much options, when dice so nice has some very amazing options can be a bother.

Because of this, i reverse engineered two other options, that i named "Custom" and "Colors". The first just removes the colorset to the faces, the second uses red and orange symbols for the Hunger and Desperation Dice.

Nothing much, shouldn't be a problem overall, been using on my private side already for a couple days!